requirements.txt
streamlit
streamlit-webrtc
numpy
av
gtts
openai
groq
python-dotenv